EFA-612 Calculate delay directly in seconds

+++ k9mail/src/main/java/security/planck/ui/passphrase/models/PassphraseState.kt
@@ -6,7 +6,7 @@ import androidx.compose.runtime.mutableStateOf
 sealed interface PassphraseState {
     object TooManyFailedAttempts : PassphraseState
     object Processing : PassphraseState
-    data class WaitAfterFailedAttempt(val seconds: Long) : PassphraseState
+    data class WaitAfterFailedAttempt(val seconds: Int) : PassphraseState
     object Success : PassphraseState
     data class CoreError(val error: Throwable?) : PassphraseState
 }
